+++ SF3Lib/Models/Structs/MPD/Model/ModelsHeader.cs
@@ -17,16 +17,16 @@ namespace SF3.Models.Structs.MPD.Model {
 
         [BulkCopy]
         [TableViewModelColumn(displayOrder: 0, isPointer: true, minWidth: 80)]
-        public int CollisionLinesHeaderOffset {
-            get => Data.GetDouble(_collisionLinesHeaderOffsetAddr);
-            set => Data.SetDouble(_collisionLinesHeaderOffsetAddr, value);
+        public uint CollisionLinesHeaderOffset {
+            get => (uint) Data.GetDouble(_collisionLinesHeaderOffsetAddr);
+            set => Data.SetDouble(_collisionLinesHeaderOffsetAddr, (int) value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(displayOrder: 1, isPointer: true, minWidth: 80)]
-        public int CollisionBlocksOffset {
-            get => Data.GetDouble(_collisionBlocksOffsetAddr);
-            set => Data.SetDouble(_collisionBlocksOffsetAddr, value);
+        public uint CollisionBlocksOffset {
+            get => (uint) Data.GetDouble(_collisionBlocksOffsetAddr);
+            set => Data.SetDouble(_collisionBlocksOffsetAddr, (int) value);
         }
 
         [BulkCopy]
